DSAlg Practice - Queue With Stacks
Implemented a queue using two seperate stacks to maintain Queue
expectations such as FIFO and what should be viewed when peeked. This
meant having to pass elements back and forth between to stacks to get to
the first element of a stack via a series of pops while maintaing
correct order when placing all of the elements back into the first
stack.